+++ include/IsoVerifierWidget.h
@@ -4,12 +4,15 @@
 
 #include <QWidget>
 
+class QFrame;
 class QLineEdit;
 class QPushButton;
+class QToolButton;
 class QTableWidget;
 class QLabel;
 class QProgressBar;
 class QTextEdit;
+class QSplitter;
 
 namespace FlashSentry {
 
@@ -29,6 +32,8 @@ public:
     void verifyMountPoint(const QString& mountPoint, const QString& deviceNode,
                           const QString& deviceLabel = {});
 
+    void refreshCatalogStatus();
+
 signals:
     void logMessageRequested(const QString& message);
     void verificationReportReady(const QString& deviceNode, const QList<IsoVerifyResult>& results);
@@ -45,6 +50,7 @@ private slots:
     void onCopyReport();
     void onUpdateCatalog();
     void onPasteTrustedHash();
+    void onScanPathEdited();
     void onVerificationFinished(const QString& location, const QString& deviceNode,
                                 const QList<IsoVerifyResult>& results);
 
@@ -52,13 +58,26 @@ private:
     void setResults(const QList<IsoVerifyResult>& results);
     void appendReport(const QString& text);
     void updateCatalogIntegrityBanner();
+    void updateMultibootBadge();
+    void updateSummaryStrip(int passed, int total, int needsSidecar);
+    void applyChromeStyles();
+    void styleResultRow(int row, bool passed);
 
     IsoVerifierWorker* m_worker = nullptr;
     QLabel* m_catalogBanner = nullptr;
+    QLabel* m_introLabel = nullptr;
+    QLabel* m_multibootBadge = nullptr;
+    QFrame* m_summaryStrip = nullptr;
+    QLabel* m_passChip = nullptr;
+    QLabel* m_failChip = nullptr;
+    QLabel* m_sidecarChip = nullptr;
     QLineEdit* m_dirEdit = nullptr;
+    QPushButton* m_verifyBtn = nullptr;
+    QToolButton* m_toolsBtn = nullptr;
     QTableWidget* m_table = nullptr;
     QLabel* m_summaryLabel = nullptr;
     QProgressBar* m_progress = nullptr;
+    QSplitter* m_splitter = nullptr;
     QTextEdit* m_reportView = nullptr;
     QList<IsoVerifyResult> m_lastResults;
     QString m_lastDeviceNode;