+++ tests/test_iso_scan_rules.cpp
@@ -1,6 +1,7 @@
 #include <QtTest>
 #include <QTemporaryDir>
 #include <QDir>
+#include <QFile>
 #include "IsoScanRules.h"
 #include "IsoVerifier.h"